Moved monitoring reponsibility from frontend to SocketServer
Added two scoped services in SocketServer to handle the monitoring of containers. They use "fire and forget" tasks, because the result is a socket event sent to the server, which means SocketServer does not need a "return value". This is achieved by the new TaskHelper static class.
Added new redux handling of receiving notifications socket events, which is quite similar to the implemented InspectData redux. Updated the depencies in the other files, that depended on the old monitoring.